Fix for compilation warnings in bitwise operations
The ADXL362 driver embedded in this header throws compiler warnings about bitwise arithmetic.  This is because there are one-liners containing multiple bitwise operators with no parens.  These are generally not recommended because they depend on operator precedence to evaluate in the correct order, and that can be quite tricky to get right.

This commit simply adds some parens that follow `C` operator precedence (doing what the statement was already doing, but explicitly).